fix(rust-sdk): accept negative creditsUsed in crawl and batch scrape status
The v2 crawl status endpoint reports creditsUsed: -1 when no billing
record exists, which is every job on a self-hosted instance without
database authentication. The SDK typed the field as Option<u32>, so
get_crawl_status, crawl, and batch_scrape failed to deserialize the
response. Widen the field to Option<i64> and add a mock regression test.

+++ apps/rust-sdk/src/batch_scrape.rs
@@ -83,8 +83,8 @@ pub struct BatchScrapeJob {
     pub completed: u32,
     /// Total number of URLs to scrape.
     pub total: u32,
-    /// Credits used by the batch scrape.
-    pub credits_used: Option<u32>,
+    /// Credits used by the batch scrape; `-1` when no billing record exists.
+    pub credits_used: Option<i64>,
     /// Expiry time of the batch data.
     pub expires_at: Option<String>,
     /// URL for the next page of results.
